pyproject.toml
[tool.poetry]
name = "memsage"
version = "0.9.0"
description = "LLM-Powered C++ Vulnerability Detection Tool"
authors = ["Gokul Krishna P <gkgokulkrishna33@gmail.com>"]
readme = "README.md"
packages = [
    {include = "memsage"}
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.4"
typer = {extras = ["all"], version = "^0.16.0"}
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
anthropic = "^0.57.1"
tqdm = "^4.67.1"
rich = "^14.0.0"
libclang = "^18.1.1"
clang = "^20.1.5"
toml = "^0.10.2"

[tool.poetry.scripts]
memsage = "memsage.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
